+++ packages/overlayscrollbars-vue/src/overlayscrollbars-vue.ts
@@ -1,2 +1,3 @@
 export { default as OverlayScrollbarsComponent } from './OverlayScrollbarsComponent.vue';
 export * from './useOverlayScrollbars';
+export * from './OverlayScrollbarsComponent.types';